requirements.txt
fastapi
uvicorn
feast[gcp] # Or your preferred Feast provider
kafka-python
pandas
pydantic